refactor: move local_files into engine/ (filesystem utility)
local_files is a filesystem I/O utility (read, write, atomic write)
used by 16 consumers across the codebase. It belongs with engine's
other infrastructure modules (shell, temp, command) rather than as
a standalone file in core/ root.

Import paths updated to crate::engine::local_files via
homeboy refactor rename.

core/ root now has 4 foundational files:
config.rs, defaults.rs, output.rs, paths.rs

776 tests pass, cargo fmt clean, zero new warnings.
