+++ src/core/config.rs
@@ -1,7 +1,7 @@
 use crate::engine::identifier;
+use crate::engine::local_files::{self, FileSystem};
 use crate::engine::text::levenshtein;
 use crate::error::Error;
-use crate::local_files::{self, FileSystem};
 use crate::output::{
     BatchResult, CreateOutput, CreateResult, MergeOutput, MergeResult, RemoveResult,
 };
