+++ src/core/extension/grammar.rs
@@ -27,8 +27,8 @@ use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::path::Path;
 
+use crate::engine::local_files;
 use crate::error::{Error, Result};
-use crate::local_files;
 
 // ============================================================================
 // Grammar definition (loaded from extension TOML/JSON)
