+++ src/core/extension/runtime_helper.rs
@@ -1,5 +1,5 @@
+use crate::engine::local_files;
 use crate::error::{Error, Result};
-use crate::local_files;
 use crate::paths;
 use std::fs;
 use std::path::PathBuf;
