+++ src/core/release/changelog/bulk.rs
@@ -4,8 +4,8 @@ use std::path::Path;
 
 use crate::component;
 use crate::config::read_json_spec_to_string;
-use crate::core::local_files::{self, FileSystem};
 use crate::core::release::version;
+use crate::engine::local_files::{self, FileSystem};
 use crate::engine::validation;
 use crate::error::{Error, Result};
 use crate::paths::resolve_path;
