+++ src/core/release/version.rs
@@ -1,11 +1,11 @@
 use crate::component::{self, Component, VersionTarget};
 use crate::config::{from_str, set_json_pointer, to_string_pretty};
 use crate::engine::hooks::{self, HookFailureMode};
+use crate::engine::local_files::{self, FileSystem};
 use crate::engine::text;
 use crate::error::{Error, Result};
 use crate::extension::{load_all_extensions, ExtensionManifest};
 use crate::is_zero;
-use crate::local_files::{self, FileSystem};
 use crate::paths::resolve_path_string;
 use crate::release::changelog;
 use regex::Regex;
